Move applyDefaults to Check
